Move interceptor trait to internal module
